requirements.txt
pafy
discord.py
opencv-python
python-dotenv